add current filters view for rating control
