+++ DemoDirectory/DemoDirectory/Demos/Sort/RelevantSort/RelevantSortDemoViewController.swift
@@ -58,99 +58,12 @@ class RelevantSortDemoViewController: UIViewController {
       stackView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
     ])
     stackView.addArrangedSubview(searchBar)
-    stackView.addArrangedSubview(relevantSortController.view)
     stackView.addArrangedSubview(statsController.label)
+    stackView.addArrangedSubview(relevantSortController.view)
     stackView.addArrangedSubview(hitsController.view)
   }
   
 }
-
-class SortByController: NSObject, SelectableSegmentController {
-  
-  let searchBar: UISearchBar
-  
-  var onClick: ((Int) -> Void)?
-  
-  init(searchBar: UISearchBar) {
-    self.searchBar = searchBar
-    super.init()
-    self.searchBar.delegate = self
-  }
-  
-  func setSelected(_ selected: Int?) {
-    searchBar.selectedScopeButtonIndex = selected ?? 0
-  }
-  
-  func setItems(items: [Int: String]) {
-    searchBar.scopeButtonTitles = items.sorted(by: \.key).map(\.value)
-  }
-
-}
-
-extension SortByController: UISearchBarDelegate {
-  
-  func searchBar(_ searchBar: UISearchBar, selectedScopeButtonIndexDidChange selectedScope: Int) {
-    onClick?(selectedScope)
-  }
-  
-}
-
-class RelevantSortToggleController: UIViewController, RelevantSortController {
-  
-  var didToggle: (() -> Void)?
-  
-  let label: UILabel
-  let button: UIButton
-  
-  init() {
-    self.label = .init()
-    self.button = .init()
-    super.init(nibName: nil, bundle: nil)
-  }
-  
-  required init?(coder: NSCoder) {
-    fatalError("init(coder:) has not been implemented")
-  }
-  
-  override func viewDidLoad() {
-    super.viewDidLoad()
-    layout()
-  }
-  
-  func layout() {
-    label.numberOfLines = 2
-    label.translatesAutoresizingMaskIntoConstraints = false
-    button.translatesAutoresizingMaskIntoConstraints = false
-    button.setTitleColor(.blue, for: .normal)
-    button.addTarget(self, action: #selector(didTapButton), for: .touchUpInside)
-    let stackView = UIStackView()
-    stackView.axis = .horizontal
-    stackView.distribution = .fillProportionally
-    stackView.translatesAutoresizingMaskIntoConstraints = false
-    stackView.addArrangedSubview(label)
-    stackView.addArrangedSubview(button)
-    view.addSubview(stackView)
-    stackView.pin(to: view, insets: .init(top: 5, left: 5, bottom: -5, right: -5))
-  }
-
-  func setItem(_ item: RelevantSortTextualRepresentation?) {
-    guard let item = item else {
-      label.text = nil
-      button.setTitle(nil, for: .normal)
-      view.isHidden = true
-      return
-    }
-    view.isHidden = false
-    label.text = item.hintText
-    button.setTitle(item.toggleTitle, for: .normal)
-  }
-  
-  @objc func didTapButton() {
-    didToggle?()
-  }
-  
-}
-
 
 class RelevantHitsController: UITableViewController, HitsController {
